For the customer!

+++ Fallborn/src/com/game/fallborn/things/Thing.java
@@ -1,7 +1,7 @@
 package com.game.fallborn.things;
 
 public class Thing {
-	public int positionX, positionY;
+	public double positionX, positionY;
 	public double xSpeed, ySpeed;
 	public double walkSpeed = 1.5;
 	public double sprintFactor = 2.0;
